refactor(ntc-data-packages): move tests to API-based integration tests
Motivation:

https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html
