+++ rust-workspace/crates/ntc-data-packages/tests/data_packages/main.rs
@@ -1,0 +1,3 @@
+//! Tests for the [`ntc_data_packages::data_packages`] API.
+
+mod json_schema;
